#38 - implemented safe author deletion with usage checks and removeAllBookLinks

+++ src/application/authors/usecases/DeleteAuthorUseCase.php
@@ -5,7 +5,10 @@ declare(strict_types=1);
 namespace app\application\authors\usecases;
 
 use app\application\authors\commands\DeleteAuthorCommand;
+use app\application\ports\AuthorUsageCheckerInterface;
 use app\application\ports\UseCaseInterface;
+use app\domain\exceptions\BusinessRuleException;
+use app\domain\exceptions\DomainErrorCode;
 use app\domain\repositories\AuthorRepositoryInterface;
 
 /**
@@ -15,6 +18,7 @@ final readonly class DeleteAuthorUseCase implements UseCaseInterface
 {
     public function __construct(
         private AuthorRepositoryInterface $authorRepository,
+        private AuthorUsageCheckerInterface $authorUsageChecker,
     ) {
     }
 
@@ -25,6 +29,15 @@ final readonly class DeleteAuthorUseCase implements UseCaseInterface
     {
         $author = $this->authorRepository->get($command->id);
 
+        if ($this->authorUsageChecker->hasSubscriptions($command->id)) {
+            throw new BusinessRuleException(DomainErrorCode::AuthorHasSubscriptions);
+        }
+
+        if ($this->authorUsageChecker->isLinkedToPublishedBooks($command->id)) {
+            throw new BusinessRuleException(DomainErrorCode::AuthorLinkedToPublishedBooks);
+        }
+
+        $this->authorRepository->removeAllBookLinks($command->id);
         $this->authorRepository->delete($author);
     }
 }
